Initial commit
This is where I'm going to base my changes from the xv6 release. I took the
current xv6 source code at the time of writing (June 1, 2016) deleted the memfs
code since I won't need it (if it is desirable to have later it will be easy to
add back in), deleted all scripts that I won't need (Promela scripts, scripts
for typesetting the book containing the source code, etc.), and removed some
unnecessary configuration options from the Makefile.

I choose to get rid of the xv6 git history rather than build on top of it for
three reasons: (1) it makes it clearer to people reading the code and browsing
the git logs where pv6 development starts, (2) this project is a significant
break from xv6, since xv6 was never focused on portability, and (3) it was just
easier to work this way.