Change build system from a flat build system to a hierarchical one.
This was tricky to get right and there are still a few rough edges. For
example, some dependencies are not properly tracked because I am not yet
sure how to get Make to do everything I want it to do. That said, the
build system works pretty well, but it may require doing a `make clean`
at times when it is not smart enough to figure out that something needs
to be updated.

I will continue tweaking the build system as the need arises, but it
provides a pretty good place to start on actual pv6 development and the
RISCV/lowRISC port currently.