Fully separate machine dependant and machine independent code
Things are coarse grained right now because that is how xv6 was
originally written. I hope to move most of the code that is currently in
the machine dependent hierarchy into the machine independent as I work
on the port. Currently many of the machine dependent files are mostly
machine independent but have a few machine dependent functions and so
have to stay where they are until I separate those parts out.